refactor: introduce app orchestration layer
