+++ lib/themes/app_theme.dart
@@ -1,0 +1,158 @@
+// Copyright 2023-2025 BenderBlog Rodriguez and contributors
+// Copyright 2025 Traintime PDA authors.
+// SPDX-License-Identifier: MPL-2.0
+
+import 'package:chinese_font_library/chinese_font_library.dart';
+import 'package:flex_color_scheme/flex_color_scheme.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+ThemeData buildLightAppTheme(FlexSchemeColor color) => FlexThemeData.light(
+  colors: color,
+  usedColors: 1,
+  surfaceMode: FlexSurfaceMode.highSurfaceLowScaffold,
+  blendLevel: 2,
+  tabBarStyle: FlexTabBarStyle.forAppBar,
+  subThemesData: const FlexSubThemesData(
+    interactionEffects: true,
+    tintedDisabledControls: true,
+    blendOnLevel: 8,
+    useM2StyleDividerInM3: true,
+    defaultRadius: 12.0,
+    elevatedButtonSchemeColor: SchemeColor.onPrimaryContainer,
+    elevatedButtonSecondarySchemeColor: SchemeColor.primaryContainer,
+    outlinedButtonOutlineSchemeColor: SchemeColor.primary,
+    toggleButtonsBorderSchemeColor: SchemeColor.primary,
+    segmentedButtonSchemeColor: SchemeColor.primary,
+    segmentedButtonBorderSchemeColor: SchemeColor.primary,
+    unselectedToggleIsColored: true,
+    sliderValueTinted: true,
+    inputDecoratorSchemeColor: SchemeColor.primary,
+    inputDecoratorIsFilled: true,
+    inputDecoratorContentPadding: EdgeInsetsDirectional.fromSTEB(
+      12,
+      16,
+      12,
+      12,
+    ),
+    inputDecoratorBackgroundAlpha: 7,
+    inputDecoratorBorderSchemeColor: SchemeColor.primary,
+    inputDecoratorBorderType: FlexInputBorderType.outline,
+    inputDecoratorRadius: 8.0,
+    inputDecoratorUnfocusedBorderIsColored: true,
+    inputDecoratorBorderWidth: 1.0,
+    inputDecoratorFocusedBorderWidth: 2.0,
+    inputDecoratorPrefixIconSchemeColor: SchemeColor.onPrimaryFixedVariant,
+    inputDecoratorSuffixIconSchemeColor: SchemeColor.primary,
+    fabUseShape: true,
+    fabAlwaysCircular: true,
+    fabSchemeColor: SchemeColor.secondary,
+    popupMenuRadius: 8.0,
+    popupMenuElevation: 3.0,
+    alignedDropdown: true,
+    dialogBackgroundSchemeColor: SchemeColor.secondaryContainer,
+    drawerIndicatorRadius: 12.0,
+    drawerIndicatorSchemeColor: SchemeColor.primary,
+    bottomNavigationBarMutedUnselectedLabel: false,
+    bottomNavigationBarMutedUnselectedIcon: false,
+    menuRadius: 8.0,
+    menuElevation: 3.0,
+    menuBarRadius: 0.0,
+    menuBarElevation: 2.0,
+    menuBarShadowColor: Color(0x00000000),
+    searchBarElevation: 1.0,
+    searchViewElevation: 1.0,
+    searchUseGlobalShape: true,
+    navigationBarSelectedLabelSchemeColor: SchemeColor.primary,
+    navigationBarSelectedIconSchemeColor: SchemeColor.onPrimary,
+    navigationBarIndicatorSchemeColor: SchemeColor.primary,
+    navigationBarIndicatorRadius: 12.0,
+    navigationRailSelectedLabelSchemeColor: SchemeColor.primary,
+    navigationRailSelectedIconSchemeColor: SchemeColor.onPrimary,
+    navigationRailUseIndicator: true,
+    navigationRailIndicatorSchemeColor: SchemeColor.primary,
+    navigationRailIndicatorOpacity: 1.00,
+    navigationRailIndicatorRadius: 12.0,
+    navigationRailBackgroundSchemeColor: SchemeColor.surface,
+    navigationRailLabelType: NavigationRailLabelType.all,
+  ),
+  keyColors: const FlexKeyColors(keepPrimary: true),
+  tones: FlexSchemeVariant.jolly.tones(Brightness.light),
+  visualDensity: FlexColorScheme.comfortablePlatformDensity,
+  cupertinoOverrideTheme: const CupertinoThemeData(applyThemeToAll: true),
+).useSystemChineseFont(Brightness.light);
+
+ThemeData buildDarkAppTheme(FlexSchemeColor color) => FlexThemeData.dark(
+  colors: color,
+  usedColors: 1,
+  surfaceMode: FlexSurfaceMode.highScaffoldLowSurface,
+  blendLevel: 2,
+  tabBarStyle: FlexTabBarStyle.forAppBar,
+  subThemesData: const FlexSubThemesData(
+    interactionEffects: true,
+    tintedDisabledControls: true,
+    blendOnLevel: 10,
+    blendOnColors: true,
+    useM2StyleDividerInM3: true,
+    defaultRadius: 12.0,
+    elevatedButtonSchemeColor: SchemeColor.onPrimaryContainer,
+    elevatedButtonSecondarySchemeColor: SchemeColor.primaryContainer,
+    outlinedButtonOutlineSchemeColor: SchemeColor.primary,
+    toggleButtonsBorderSchemeColor: SchemeColor.primary,
+    segmentedButtonSchemeColor: SchemeColor.primary,
+    segmentedButtonBorderSchemeColor: SchemeColor.primary,
+    unselectedToggleIsColored: true,
+    sliderValueTinted: true,
+    inputDecoratorSchemeColor: SchemeColor.primary,
+    inputDecoratorIsFilled: true,
+    inputDecoratorContentPadding: EdgeInsetsDirectional.fromSTEB(
+      12,
+      16,
+      12,
+      12,
+    ),
+    inputDecoratorBackgroundAlpha: 40,
+    inputDecoratorBorderSchemeColor: SchemeColor.primary,
+    inputDecoratorBorderType: FlexInputBorderType.outline,
+    inputDecoratorRadius: 8.0,
+    inputDecoratorUnfocusedBorderIsColored: true,
+    inputDecoratorBorderWidth: 1.0,
+    inputDecoratorFocusedBorderWidth: 2.0,
+    inputDecoratorPrefixIconSchemeColor: SchemeColor.primaryFixed,
+    inputDecoratorSuffixIconSchemeColor: SchemeColor.primary,
+    fabUseShape: true,
+    fabAlwaysCircular: true,
+    fabSchemeColor: SchemeColor.secondary,
+    popupMenuRadius: 8.0,
+    popupMenuElevation: 3.0,
+    alignedDropdown: true,
+    drawerIndicatorRadius: 12.0,
+    drawerIndicatorSchemeColor: SchemeColor.primary,
+    bottomNavigationBarMutedUnselectedLabel: false,
+    bottomNavigationBarMutedUnselectedIcon: false,
+    menuRadius: 8.0,
+    menuElevation: 3.0,
+    menuBarRadius: 0.0,
+    menuBarElevation: 2.0,
+    menuBarShadowColor: Color(0x00000000),
+    searchBarElevation: 1.0,
+    searchViewElevation: 1.0,
+    searchUseGlobalShape: true,
+    navigationBarSelectedLabelSchemeColor: SchemeColor.primary,
+    navigationBarSelectedIconSchemeColor: SchemeColor.onPrimary,
+    navigationBarIndicatorSchemeColor: SchemeColor.primary,
+    navigationBarIndicatorRadius: 12.0,
+    navigationRailSelectedLabelSchemeColor: SchemeColor.primary,
+    navigationRailSelectedIconSchemeColor: SchemeColor.onPrimary,
+    navigationRailUseIndicator: true,
+    navigationRailIndicatorSchemeColor: SchemeColor.primary,
+    navigationRailIndicatorOpacity: 1.00,
+    navigationRailIndicatorRadius: 12.0,
+    navigationRailBackgroundSchemeColor: SchemeColor.surface,
+    navigationRailLabelType: NavigationRailLabelType.all,
+  ),
+  keyColors: const FlexKeyColors(),
+  tones: FlexSchemeVariant.jolly.tones(Brightness.dark),
+  visualDensity: FlexColorScheme.comfortablePlatformDensity,
+  cupertinoOverrideTheme: const CupertinoThemeData(applyThemeToAll: true),
+).useSystemChineseFont(Brightness.dark);
